Crate ic_protobuf

Modules

crypto

log

messaging

p2p

proxy
A proxy for encoding arbitrary Rust structs in protocol buffer binary format by means of a "proxy message" (a generated prost::Message) plus Into / TryInto implementations to convert between the Rust struct and the proxy message.

registry

state

types

Macros

import_mod
Declare a sub-module of the log module

serialize_fallback_for
Emit the given LogEntry field in LogEntry.serialize_fallback