Module ic_protobuf::proxy

A proxy for encoding arbitrary Rust structs in protocol buffer binary format by means of a "proxy message" (a generated prost::Message) plus Into / TryInto implementations to convert between the Rust struct and the proxy message.

The only major downside when compared to serde (which doesn't support protocol buffers) or using a prost-generated struct directly is that (due to prost-generated structs insisting on owning their data, e.g. using Vec<u8> instead of &[u8]) is that encoding consumes the provided struct.

Example

use pb;  // Module containing generated `prost::Message`.

// Rust struct to be encoded/decoded.
struct Block {
    pub height: Height,
    pub contents: Blob,
}

// To be provided by the client.
impl From<Block> for pb::Block {
    ...
}
impl TryFrom<pb::Block> for Block {
    ...
}

let b = Block {
    height: Height(1),
    payload: Blob(vec![1, 2, 3]),
};

// Encode.
let bytes: Vec<u8> = pb::Block::proxy_encode(b).unwrap();

// Decode.
assert_eq!(
    Block {
        height: Height(1),
        payload: Blob(vec![1, 2, 3]),
    },
    pb::Block::proxy_decode(&bytes).unwrap()
);

Enums

ProxyDecodeError
Errors that may result when mapping a proto to a Rust struct.

Traits

ProtoProxy
Allows a prost::Message to act as a proxy for encoding and decoding a Rust type T, as long as Into / TryInto implementations are provided to convert from one to the other and back.

Functions

try_from_option_field
Converts an optional proto field into a Rust type using From / TryFrom. Returns Err(ProtoMappingError::MissingField(field)) if the field is None.