Module ic_protobuf::registry::subnet::v1

Structs

CatchUpPackageContents
Contains the initial DKG transcripts for the subnet and materials to construct a base CUP (i.e. a CUP with no dependencies on previous CUPs or blocks). Such CUP materials can be used to construct the genesis CUP or a recovery CUP in the event of a subnet stall.

DkgId
The dkg id

GossipConfig
Per subnet P2P configuration Note: protoc is mangling the name P2PConfig to P2pConfig

InitialDkgTranscriptRecord

InitialNiDkgTranscriptRecord
Initial non-interactive DKG transcript record

RegistryStoreUri

SubnetListRecord
A list of subnet ids of all subnets present in this instance of the IC.

SubnetRecord
A subnet: A logical group of nodes that run consensus

Enums

SubnetType
Represents the type of subnet. Subnets of different type might exhibit different behavior, e.g. being more restrictive in what operations are allowed or privileged compared to other subnet types.